dev-toolkits - remove language sdks; add devbox
The sheer hairiness of managing SDKs is pretty much why I created the
`dev-toolkits` package in the first place.

`devbox` takes a lot of the pain out of that management by enabling us
to push that responsibility off on Nix, which handles versioned
dependency chains much more cleanly, even on Arch Linux.